requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==3.0.1
email-validator==2.1.0
reportlab==4.0.7
openai==1.6.1
requests==2.31.0
gunicorn==21.2.0
Authlib==1.3.0
